==> notebooks/10a-twitter-follow-community.ipynb <==
import os
import logging

import psycopg2
import community
import numpy as np
import pandas as pd
import networkx as nx

import matplotlib as mp
%matplotlib inline
import matplotlib.pyplot as plt

from IPython.display import display

from tqdm.notebook import tqdm

from nltk.corpus import stopwords as sw

from pandasql import sqldf
pysqldf = lambda q: sqldf(q, globals())

import utils as ut

fmt = '%(asctime)s : %(levelname)s : %(message)s'
logging.basicConfig(format=fmt, level=logging.INFO)
logger = logging.getLogger(__name__)

pd.options.display.max_rows = 100
pd.options.display.max_colwidth = None

os.chdir(os.path.expanduser('~/github/masthesis/'))

os.makedirs('data/twitter/community/', exist_ok=True)

# Load data

### User profile data

user_data = pd.read_csv('data/samples/twitter/user-data.csv', index_col='user_id')
display(user_data.shape)

user_show = pd.read_csv('data/samples/twitter/user-show.csv')
display(user_show.shape)

### Follow graph

follow_graph = pd.read_csv('data/samples/twitter/follow-graph.csv')
follow_graph = follow_graph.rename({'source_user_id': 'source', 'target_user_id': 'target'}, axis=1)

fg = nx.from_pandas_edgelist(follow_graph, source='source', target='target',
                             edge_attr=None, create_using=nx.DiGraph())

display(follow_graph.shape)

follow_graph.head()

fg.order()

fg.size()

# Confirm connectivity

ucomps = [x for x in nx.connected_components(fg.to_undirected())]
display(len(ucomps))

scomps = [x for x in nx.strongly_connected_components(fg)]
display(len(scomps))

pd.Series([len(x) for x in scomps]).value_counts()

# Community detection + centrality

## Graph as-is

partition = community.best_partition(fg.to_undirected())
partition = pd.Series(partition, name='community')

comms = pd.DataFrame(partition).merge(user_data[['screen_name', 'name', 'description', 'location', 'is_radio']],
                                      how='inner', left_index=True, right_index=True)
comms.index.name = 'user_id'

comms.community.value_counts()

comms.loc[comms.is_radio, 'community'].value_counts()

comms.sample(frac=1).groupby('community').head(10).sort_values('community')

## Quotient graph collapsing shows

The `nx.quotient_graph()` function is very, very slow...

qg = fg.to_undirected().copy()

for sh in user_show.show_id.unique():
    nodes = {i for i in user_show.loc[user_show.show_id == sh, 'user_id']}
    
    rep = list(nodes)[0]
    for n in nodes - {rep}:
        if n not in qg.nodes: # has no followers
            continue
        
        edges = list(qg.edges(n))
        for e1, e2 in edges:
            assert e1 == n

            if e2 == rep:
                continue
            else:
                qg.add_edge(rep, e2)
        
        qg.remove_node(n)

qpartition = community.best_partition(qg)
qpartition = pd.Series(qpartition, name='community')

qcomms = pd.DataFrame(qpartition).merge(user_data[['screen_name', 'name', 'description', 'location', 'is_radio']],
                                        how='inner', left_index=True, right_index=True)
qcomms.index.name = 'user_id'

qcomms.community.value_counts()

qcomms.loc[qcomms.is_radio, 'community'].value_counts()

qcomms.sample(frac=1).groupby('community').head(10).sort_values('community')

## Betweenness centrality

bc = pd.Series(nx.betweenness_centrality(fg), name='centrality')
bc = pd.DataFrame(bc).reset_index().rename({'index': 'user_id'}, axis=1).set_index('user_id')

qbc = pd.Series(nx.betweenness_centrality(qg), name='centrality')
qbc = pd.DataFrame(qbc).reset_index().rename({'index': 'user_id'}, axis=1).set_index('user_id')

# Match to shows

## Graph as-is

rcomms = comms.merge(user_show, how='inner', right_on='user_id', left_index=True)

rcomms.community.value_counts()

dupes = rcomms.groupby('show_id').community.nunique().sort_values()
dupes = dupes[dupes > 1].index.tolist()

rcomms.loc[rcomms.show_id.isin(dupes), ['show_id', 'show_name']].drop_duplicates()

show_community = rcomms.loc[~rcomms.show_id.isin(dupes), ['show_id', 'community']]
show_community = show_community.reset_index().drop('index', axis=1).set_index('show_id').community

show_community.value_counts()

## Quotient graph

rqcomms = qcomms.merge(user_show, how='inner', right_on='user_id', left_index=True)

rqcomms.community.value_counts()

dupes = rqcomms.groupby('show_id').community.nunique().sort_values()
dupes = dupes[dupes > 1].index.tolist()

rcomms.loc[rcomms.show_id.isin(dupes), ['show_id', 'show_name']].drop_duplicates()

qshow_community = rqcomms.loc[~rqcomms.show_id.isin(dupes), ['show_id', 'community']]
qshow_community = qshow_community.reset_index().drop('index', axis=1).set_index('show_id').community

qshow_community.value_counts()

## Betweenness centrality

bc_show = qbc.merge(user_show, how='inner', right_on='user_id', left_index=True)
bc_show = bc_show[['show_id', 'centrality']].set_index('show_id')

bc_show.head()

bc_show.centrality.hist(bins=50)

# Write out data

nx.write_edgelist(qg, path='data/twitter/community/quotient-follow-graph.csv', delimiter=',')

comms[['community']].to_csv('data/twitter/community/follow-community-all.csv')

qcomms[['community']].to_csv('data/twitter/community/quotient-follow-community-all.csv')

bc.to_csv('data/twitter/community/follow-centrality.csv')

qbc.to_csv('data/twitter/community/quotient-follow-centrality.csv')

show_community.to_csv('data/twitter/community/follow-community-by-show.csv')

qshow_community.to_csv('data/twitter/community/quotient-follow-community-by-show.csv')

bc_show.to_csv('data/twitter/community/quotient-follow-centrality-by-show.csv')
