==> notebooks/11b-dataset-stats-firehose.ipynb <==
import os
import json
import gzip
import random
import logging

import numpy as np
import pandas as pd

logger = logging.getLogger(__name__)

fmt = '%(asctime)s : %(levelname)s : %(message)s'
logging.basicConfig(format=fmt, level=logging.INFO)

os.chdir(os.path.expanduser('~/github/masthesis/'))

seed = 2969591811

random.seed(seed)
np.random.seed(seed)

# Manual - decahose

manual = pd.read_csv('data/paper-round-3/event-annotated/raw-decahose.csv')

manual.shape[0]

# Automatic - decahose

with gzip.open('data/paper-round-3/event-annotated/auto-sample.csv.gz', 'rt') as f:
    automatic = pd.read_csv(f)
    automatic.drop(automatic.loc[automatic['year'] > 2021].index, axis=0, inplace=True)

decahose_ids = automatic \
    .loc[automatic['kind'] == 'decahose', 'id'] \
    .str.replace('D', '') \
    .astype(int)

decahose_ids.nunique()

# Total - decahose

len(set(manual['id'].tolist() + decahose_ids.tolist()))

# Total

elite_ids = automatic \
    .loc[automatic['kind'] == 'elite', 'id'] \
    .str.replace('E', '') \
    .astype(int)

len(set(manual['id'].tolist() + decahose_ids.tolist() + elite_ids.tolist()))
